pyproject.toml
[project]
name = "ipaddress-su"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "attrs>=25.3.0",
    "django>=5.1.6",
    "django-environ>=0.12.0",
    "django-ipware>=7.0.1",
    "django-tinymce>=4.1.0",
    "dnspython>=2.7.0",
    "gunicorn>=23.0.0",
    "httpx[http2]>=0.28.1",
    "markdown>=3.7",
    "netaddr>=1.3.0",
    "pillow>=11.1.0",
    "psycopg>=3.2.9",
    "pytest-asyncio>=1.0.0",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.1.0",
    "starlette>=0.45.3",
    "uuid6>=2024.7.10",
    "uvicorn>=0.34.0",
    "validators>=0.34.0",
]

# -- Example FILE: pyproject.toml
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
# -- recommended but optional:
python_files = ["test_*.py", "*_test.py", "testing/python/*.py", "*/tests*.py"]

[tool.ruff]
line-length = 79
exclude = [".venv"]

[tool.ruff.format]
quote-style = "single"
#indent-style = "tab"
#docstring-code-format = true
